refactor(elements): Consolidate SignIn Reset Password machine into Router
