+++ .changeset/silly-pillows-count.md
@@ -1,0 +1,5 @@
+---
+"@clerk/elements": patch
+---
+
+Consolidate SignIn Reset Password machine with Router
