+++ packages/elements/src/internals/machines/sign-in/index.ts
@@ -1,13 +1,10 @@
 export { SignInFirstFactorMachine, SignInSecondFactorMachine } from './verification.machine';
 export { SignInRouterMachine, SignInRouterMachineId } from './router.machine';
-export { SignInResetPasswordMachine, SignInResetPasswordMachineId } from './reset-password.machine';
 
 export { SignInSafeIdentifierSelectorForStrategy, SignInSalutationSelector } from './router.selectors';
 
 export type { TSignInRouterMachine } from './router.machine';
 export type { TSignInFirstFactorMachine, TSignInSecondFactorMachine } from './verification.machine';
-export type { TSignInResetPasswordMachine } from './reset-password.machine';
 
 export * from './verification.types';
 export * from './router.types';
-export * from './reset-password.types';
