+++ packages/elements/src/react/sign-in/context/router.context.ts
@@ -2,7 +2,6 @@ import type { ActorRefFrom, AnyActorRef, AnyStateMachine, SnapshotFrom } from 'x
 
 import type {
   TSignInFirstFactorMachine,
-  TSignInResetPasswordMachine,
   TSignInRouterMachine,
   TSignInSecondFactorMachine,
 } from '~/internals/machines/sign-in';
@@ -18,7 +17,6 @@ function useSignInStep<M extends AnyStateMachine, T = ActorRefFrom<M>>(name: str
 
 export const useSignInFirstFactorStep = () => useSignInStep<TSignInFirstFactorMachine>('firstFactor');
 export const useSignInSecondFactorStep = () => useSignInStep<TSignInSecondFactorMachine>('secondFactor');
-export const useSignInResetPasswordStep = () => useSignInStep<TSignInResetPasswordMachine>('resetPassword');
 
 export const useSignInPasskeyAutofill = () =>
   SignInRouterCtx.useSelector(state => state.context.webAuthnAutofillSupport);
